backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.5",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "httpx>=0.28",
    "Pillow>=11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
